+++ src/ServerApi/Adnc.sln
@@ -65,8 +65,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Adnc.Ord.WebApi", "Services
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Adnc.Infra.Common", "Infrastructures\Adnc.Infra.Common\Adnc.Infra.Common.csproj", "{3BCA661E-0A8D-44E7-9FF0-437F3561D050}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Adnc.Infra.Mq", "Infrastructures\Adnc.Infra.Mq\Adnc.Infra.Mq.csproj", "{56AE1902-C1BC-4617-8E76-A1614E5CDD27}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Adnc.Infra.Consul", "Infrastructures\Adnc.Infra.Consul\Adnc.Infra.Consul.csproj", "{45A76D25-58EB-4AE0-85B4-31F7B777CA15}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Adnc.Infra.EventBus", "Infrastructures\Adnc.Infra.EventBus\Adnc.Infra.EventBus.csproj", "{13A88A12-58E6-403D-9BC0-66A04C36BABA}"
@@ -291,14 +289,6 @@ Global
 		{3BCA661E-0A8D-44E7-9FF0-437F3561D050}.Release|Any CPU.Build.0 = Release|Any CPU
 		{3BCA661E-0A8D-44E7-9FF0-437F3561D050}.Release|x64.ActiveCfg = Release|Any CPU
 		{3BCA661E-0A8D-44E7-9FF0-437F3561D050}.Release|x64.Build.0 = Release|Any CPU
-		{56AE1902-C1BC-4617-8E76-A1614E5CDD27}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{56AE1902-C1BC-4617-8E76-A1614E5CDD27}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{56AE1902-C1BC-4617-8E76-A1614E5CDD27}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{56AE1902-C1BC-4617-8E76-A1614E5CDD27}.Debug|x64.Build.0 = Debug|Any CPU
-		{56AE1902-C1BC-4617-8E76-A1614E5CDD27}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{56AE1902-C1BC-4617-8E76-A1614E5CDD27}.Release|Any CPU.Build.0 = Release|Any CPU
-		{56AE1902-C1BC-4617-8E76-A1614E5CDD27}.Release|x64.ActiveCfg = Release|Any CPU
-		{56AE1902-C1BC-4617-8E76-A1614E5CDD27}.Release|x64.Build.0 = Release|Any CPU
 		{45A76D25-58EB-4AE0-85B4-31F7B777CA15}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{45A76D25-58EB-4AE0-85B4-31F7B777CA15}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{45A76D25-58EB-4AE0-85B4-31F7B777CA15}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -460,7 +450,6 @@ Global
 		{65A612B2-CF1F-4EFC-BC4E-32AF75FAFC37} = {6659C0D2-F08D-40B3-99ED-AA85505A5DA7}
 		{660C5BA1-B2F6-45CA-8EF0-CC760F0A4C0B} = {6659C0D2-F08D-40B3-99ED-AA85505A5DA7}
 		{3BCA661E-0A8D-44E7-9FF0-437F3561D050} = {3B899133-BEAC-4438-977F-B28FDFF3AE27}
-		{56AE1902-C1BC-4617-8E76-A1614E5CDD27} = {3B899133-BEAC-4438-977F-B28FDFF3AE27}
 		{45A76D25-58EB-4AE0-85B4-31F7B777CA15} = {3B899133-BEAC-4438-977F-B28FDFF3AE27}
 		{13A88A12-58E6-403D-9BC0-66A04C36BABA} = {3B899133-BEAC-4438-977F-B28FDFF3AE27}
 		{43D81943-1D5A-4175-89BA-B48B486D1AA6} = {3B899133-BEAC-4438-977F-B28FDFF3AE27}
