+++ src/ServerApi/Infrastructures/Adnc.Infra.EventBus/Adnc.Infra.EventBus.csproj
@@ -12,6 +12,12 @@
 		<Copyright>Alpha</Copyright>
 	</PropertyGroup>
 	<ItemGroup>
-	  <PackageReference Include="DotNetCore.CAP" Version="3.0.4" />
+		<PackageReference Include="Autofac" Version="6.0.0" />
+		<PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="3.1.9" />
+		<PackageReference Include="Microsoft.Extensions.Options" Version="3.1.9" />
+		<PackageReference Include="Polly" Version="7.2.1" />
+		<PackageReference Include="RabbitMQ.Client" Version="5.1.0" />
+		<PackageReference Include="System.Text.Json" Version="4.7.2" />
+	   <PackageReference Include="DotNetCore.CAP" Version="3.0.4" />
 	</ItemGroup>
 </Project>
