+++ src/ServerApi/Infrastructures/Adnc.Infra.EventBus/Cap/EbConsts.cs
@@ -1,4 +1,4 @@
-﻿namespace Adnc.Infra.EventBus
+﻿namespace Adnc.Infra.EventBus.Cap
 {
     public class EbConsts
     {
