+++ src/ServerApi/Infrastructures/Adnc.Infra.EventBus/Cap/NullCapPublisher.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Adnc.Infra.EventBus
+namespace Adnc.Infra.EventBus.Cap
 {
     /// <summary>
     /// 空的Cap发布者，给单元测试用
