--- src/ServerApi/Infrastructures/Adnc.Infra.Mq/Adnc.Infra.Mq.csproj
@@ -1,24 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<PropertyGroup>
-		<TargetFramework>netstandard2.1</TargetFramework>
-		<AssemblyName>Adnc.Infra.Mq</AssemblyName>
-		<RootNamespace>Adnc.Infra.Mq</RootNamespace>
-		<Version>0.8.1</Version>
-		<Authors>Alpha.Yu</Authors>
-		<Product>Adnc</Product>
-		<RepositoryUrl>https://github.com/AlphaYu</RepositoryUrl>
-		<RepositoryType>git</RepositoryType>
-		<PackageProjectUrl>https://www.aspdotnetcore.net</PackageProjectUrl>
-		<Copyright>Alpha</Copyright>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<PackageReference Include="Autofac" Version="6.0.0" />
-		<PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="3.1.9" />
-		<PackageReference Include="Microsoft.Extensions.Options" Version="3.1.9" />
-		<PackageReference Include="Polly" Version="7.2.1" />
-		<PackageReference Include="RabbitMQ.Client" Version="5.1.0" />
-		<PackageReference Include="System.Text.Json" Version="4.7.2" />
-	</ItemGroup>
-</Project>
