+++ src/ServerApi/Services/Adnc.Maint/Adnc.Maint.Application/EventSubscribers/LoginLogMqConsumer.cs
@@ -5,26 +5,26 @@ using System.Collections.Generic;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Microsoft.Extensions.DependencyInjection;
+using Adnc.Maint.Core.Entities;
 using Adnc.Core.Shared.IRepositories;
-using Adnc.Infra.Mq.RabbitMq;
-using Adnc.Core.Maint.Entities;
+using Adnc.Infra.EventBus.RabbitMq;
 using Adnc.Maint.Application.Contracts.Consts;
 
-namespace Adnc.Maint.Application.Mq
+namespace Adnc.Maint.Application.EventSubscribers
 {
     /// <summary>
-    /// 操作日志消费者
+    /// 登录日志消费者
     /// </summary>
-    public sealed class OpsLogMqConsumer : BaseRabbitMqConsumer
+    public sealed class LoginLogMqConsumer : BaseRabbitMqConsumer
     {
         // 因为Process函数是委托回调,直接将其他Service注入的话两者不在一个scope,
         // 这里要调用其他的Service实例只能用IServiceProvider CreateScope后获取实例对象
         private readonly IServiceProvider _services;
-        private readonly ILogger<OpsLogMqConsumer> _logger;
+        private readonly ILogger<LoginLogMqConsumer> _logger;
 
-        public OpsLogMqConsumer(IOptionsSnapshot<RabbitMqConfig> options
-           , ILogger<OpsLogMqConsumer> logger
-           , IServiceProvider services) 
+        public LoginLogMqConsumer(IOptionsSnapshot<RabbitMqConfig> options
+           , ILogger<LoginLogMqConsumer> logger
+           , IServiceProvider services)
             : base(options, logger)
         {
             _services = services;
@@ -53,7 +53,7 @@ namespace Adnc.Maint.Application.Mq
         /// <returns></returns>
         protected override string[] GetRoutingKeys()
         {
-            return new[] { MqRoutingKeys.OpsLog };
+            return new[] { MqRoutingKeys.Loginlog }; 
         }
 
         /// <summary>
@@ -63,7 +63,8 @@ namespace Adnc.Maint.Application.Mq
         protected override QueueConfig GetQueueConfig()
         {
             var config = GetCommonQueueConfig();
-            config.Name = "q-adnc-maint-opslog";
+
+            config.Name = "q-adnc-maint-loginlog";
             config.AutoAck = false;
             config.PrefetchCount = 5;
             config.Arguments = new Dictionary<string, object>()
@@ -71,7 +72,7 @@ namespace Adnc.Maint.Application.Mq
                      //设置当前队列的DLX
                     { "x-dead-letter-exchange",MqExchanges.Dead} 
                     //设置DLX的路由key，DLX会根据该值去找到死信消息存放的队列
-                    ,{ "x-dead-letter-routing-key",MqRoutingKeys.OpsLog}
+                    ,{ "x-dead-letter-routing-key",MqRoutingKeys.Loginlog}
                     //设置消息的存活时间，即过期时间(毫秒)
                     ,{ "x-message-ttl",1000*60}
                   };
@@ -92,8 +93,8 @@ namespace Adnc.Maint.Application.Mq
             {
                 using (var scope = _services.CreateScope())
                 {
-                    var repository = scope.ServiceProvider.GetRequiredService<IMongoRepository<SysOperationLog>>();
-                    var entity = JsonSerializer.Deserialize<SysOperationLog>(message);
+                    var repository = scope.ServiceProvider.GetRequiredService<IMongoRepository<SysLoginLog>>();
+                    var entity = JsonSerializer.Deserialize<SysLoginLog>(message);
                     await repository.AddAsync(entity);
                     result = true;
                 }
