--- src/ServerApi/Services/Adnc.Usr/Adnc.Usr.Core.Shared/Adnc.Usr.Core.Shared/Adnc.Usr.Core.Shared.csproj
@@ -1,7 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.1</TargetFramework>
-  </PropertyGroup>
-
-</Project>
