--- src/ServerApi/Services/Adnc.Usr/Adnc.Usr.Core.Shared/Adnc.Usr.Core.Shared/EntityConsts.cs
@@ -1,21 +1,0 @@
-﻿using System;
-
-namespace Adnc.Usr.Core.Entities.Consts
-{
-    public static class UserConsts
-    {
-        public const int Account_MaxLength = 16;
-        public const int Avatar_MaxLength = 64;
-        public const int Email_Maxlength = 32;
-        public const int Name_Maxlength = 16;
-        public const int Password_Maxlength = 32;
-        public const int Phone_Maxlength = 11;
-        public const int RoleIds_Maxlength = 72;
-        public const int Salt_Maxlength = 6;
-    }
-
-    public static class RoleConsts
-    {
-
-    }
-}
