+++ src/ServerApi/Services/Shared/Adnc.Application.Shared/Adnc.Application.Shared.csproj
@@ -28,7 +28,6 @@
 	  <ProjectReference Include="..\..\..\Infrastructures\Adnc.Infra.Common\Adnc.Infra.Common.csproj" />
 	  <ProjectReference Include="..\..\..\Infrastructures\Adnc.Infra.Caching\Adnc.Infra.Caching.csproj" />
 	  <ProjectReference Include="..\..\..\Infrastructures\Adnc.Infra.Mapper\Adnc.Infra.Mapper.csproj" />
-	  <ProjectReference Include="..\..\..\Infrastructures\Adnc.Infra.Mq\Adnc.Infra.Mq.csproj" />
 	  <ProjectReference Include="..\Adnc.Core.Shared\Adnc.Core.Shared.csproj" />
 	</ItemGroup>
