+++ src/ServerApi/Services/Shared/Adnc.Core.Shared/AdncCoreModule.cs
@@ -4,6 +4,7 @@ using System.Reflection;
 using Adnc.Core.Shared.Entities;
 using Adnc.Core.Shared.Interceptors;
 using Adnc.Infra.EventBus;
+using Adnc.Infra.EventBus.Cap;
 using Autofac;
 using Autofac.Extras.DynamicProxy;
